+++ include/effects/MovingEffect.h
@@ -10,14 +10,14 @@ class MovingEffect : public WrapperEffect {
 private:
     float ledsPerSecond;
     float currentPosition;
-    int startPosition;
-    int endPosition;
+    float startPosition;
+    float endPosition;
     float elapsedPercentage;
     unsigned long startTimeMillis;
 
 public:
-    MovingEffect(BaseLedPipelineStage *stage, float ledsPerSecond, int startPosition = 0,
-                 int endPosition = TemporaryLedData::size);
+    MovingEffect(BaseLedPipelineStage *stage, float ledsPerSecond, float startPosition = 0,
+                 float endPosition = TemporaryLedData::size);
 
     void calculate(float startIndex, TemporaryLedData &tempData) override;
