+++ src/effects/AbsolutePositionEffect.cpp
@@ -3,7 +3,7 @@
 using namespace ledpipelines;
 using namespace ledpipelines::effects;
 
-AbsolutePositionEffect::AbsolutePositionEffect(BaseLedPipelineStage *stage, int startPosition)
+AbsolutePositionEffect::AbsolutePositionEffect(BaseLedPipelineStage *stage, float startPosition)
         : WrapperEffect(stage), startPosition(startPosition) {}
 
 void AbsolutePositionEffect::calculate(float startIndex, TemporaryLedData &tempData) {
